Steve's requests:
Would you be so kind as to change the color of the banner message, "Offering family...meaningful prayer.", to 7D4900   ? Thanks

bigactual [5:54 PM]
When you shrink the screen width in chrome on Windows, About the Author creeps onto the picture of Susan

bigactual [5:55 PM]
could you add Susan's whole name into the About the Author Excerpt on the home page?  Please/thanks.

bigactual [5:55 PM]
Please change the font in the address footer to the same font in the slogan, "Teach your children . . ."
bigactual
[5:58 PM]
Pasted image at 2015-05-26, 6:58 PM

245KB PNG shared with you • 1 comment • Open original
About the Author links to the archive of her books. Please re-route that hyper link to an anchor on the About page. Successful navigation would then look like this:
bigactual
[6:52 PM]
Pasted image at 2015-05-26, 7:52 PM

3MB PNG shared with you • 1 comment • Open original
I'm adding the other books to the site and the archive now looks like this:

bigactual [6:53 PM]
The Site Nav bar is missing and the archive list is only showing the most recent. Need to make sure all books are showing.

bigactual [10:39 PM]
Testimonial Pictures not available for certain individuals. Require default image to keep layout in check. Logo?